Usuarios/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
sqlalchemy==2.0.19
psycopg2-binary==2.9.6
python-jose==3.3.0
passlib==1.7.4
python-dotenv==0.21.1
bcrypt==4.0.1
pydantic==2.1.1
email-validator==2.0.0.post2
